+++ packages/cli/test/__mocks__/fs-extra.js
@@ -14,7 +14,9 @@ fs.__toJSON = () => vol.toJSON();
 
 fs.copy = (...args) => vol.promises.copyFile(...args);
 fs.ensureDir = (dir) => vol.promises.mkdir(dir, { recursive: true });
+fs.existsSync = (...args) => vol.existsSync(...args);
 fs.pathExists = (...args) => Promise.resolve(vol.existsSync(...args));
 fs.readFile = (...args) => vol.promises.readFile(...args);
+fs.writeFile = (...args) => vol.promises.writeFile(...args);
 
 module.exports = fs;